requirements.txt
fastapi
uvicorn
pandas
numpy
sentence-transformers
faiss-cpu
scikit-learn
